backend/app/schemas/auth.py
from typing import Optional
from pydantic import BaseModel, EmailStr
from uuid import UUID

class Token(BaseModel):
    access_token: str
    token_type: str

class TokenPayload(BaseModel):
    sub: Optional[str] = None

class UserBase(BaseModel):
    email: Optional[EmailStr] = None
    full_name: Optional[str] = None
    is_active: bool = True

class UserCreate(UserBase):
    email: EmailStr
    password: str
    tenant_id: UUID

class UserUpdate(UserBase):
    password: Optional[str] = None

class UserOut(UserBase):
    id: UUID
    tenant_id: UUID
    role: str

    class Config:
        from_attributes = True
