backend/requirements.txt
fastapi[all]
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
celery
redis
alembic
psycopg2-binary
python-dotenv
httpx
slowapi
